Move schedule constants into function
